+++ packages/shared/src/components/cards/FeedItemContainer.tsx
@@ -15,7 +15,7 @@ import {
 } from './RaisedLabel';
 import ConditionalWrapper from '../ConditionalWrapper';
 import { useFeedPreviewMode } from '../../hooks';
-import useBookmarkProvider from '../../hooks/useBookmarkProvider';
+import useBookmarkProvider from '../../hooks/bookmark/useBookmarkProvider';
 
 export interface FlagProps extends Pick<Post, 'trending' | 'pinnedAt'> {
   listMode?: boolean;
