+++ packages/shared/src/hooks/bookmark/common.ts
@@ -1,0 +1,3 @@
+export interface UseBookmarkProviderProps {
+  bookmarked: boolean;
+}
