+++ packages/shared/src/hooks/bookmark/index.ts
@@ -1,0 +1,1 @@
+export * from './useJustBookmarked';
