+++ packages/shared/src/hooks/bookmark/useJustBookmarked.ts
@@ -1,15 +1,6 @@
 import { useEffect, useRef } from 'react';
-import { useActiveFeedNameContext } from '../contexts';
-import { SharedFeedPage } from '../components/utilities';
-import { isNullOrUndefined } from '../lib/func';
-
-interface UseBookmarkProviderReturn {
-  highlightBookmarkedPost: boolean;
-}
-
-interface UseBookmarkProviderProps {
-  bookmarked: boolean;
-}
+import { isNullOrUndefined } from '../../lib/func';
+import { UseBookmarkProviderProps } from './common';
 
 interface UseJustBookmarked {
   justBookmarked: boolean;
@@ -38,20 +29,3 @@ export const useJustBookmarked = ({
 
   return { justBookmarked, wasBookmarked: wasBookmarkedRef.current };
 };
-
-const useBookmarkProvider = ({
-  bookmarked = false,
-}: UseBookmarkProviderProps): UseBookmarkProviderReturn => {
-  const { feedName } = useActiveFeedNameContext();
-  const { justBookmarked, wasBookmarked } = useJustBookmarked({ bookmarked });
-
-  const isMyFeed = feedName === SharedFeedPage.MyFeed;
-  const highlightBookmarkedPost =
-    isMyFeed && wasBookmarked && !justBookmarked && bookmarked;
-
-  return {
-    highlightBookmarkedPost,
-  };
-};
-
-export default useBookmarkProvider;
